+++ pkg/action/cl001/ac009/explainer.go
@@ -5,5 +5,19 @@ import (
 )
 
 func (e *Explainer) explain(ctx context.Context) (string, error) {
-	return "", nil
+	s := `
+Check that all relevant CRs of the tenant cluster management got properly
+cleaned up eventually during the transition of cluster deletion. This check
+considers the following CRs.
+
+	* Cluster
+	* AWSCluster
+	* G8sControlPlane
+	* AWSControlPlane
+	* MachineDeployment
+	* AWSMachineDeployment
+
+`
+
+	return s, nil
 }
